Migrate from buildSrc to includeBuild
