--- buildSrc/build.gradle
@@ -1,4 +1,0 @@
-apply plugin: "groovy"
-
-sourceSets.main.groovy.srcDirs = ["../spock-gradle/src/main/groovy"]
-sourceSets.main.resources.srcDirs = ["../spock-gradle/src/main/resources"]
