+++ settings.gradle
@@ -10,6 +10,8 @@ pluginManagement {
     id "com.github.ben-manes.versions" version "0.39.0"
     id "biz.aQute.bnd.builder" version "5.3.0"
   }
+
+  includeBuild "spock-gradle"
 }
 
 plugins {
@@ -91,7 +93,6 @@ if (JavaVersion.current().isJava8()) {
 }
 
 include "spock-unitils"
-include "spock-gradle"
 
 //include "spock-spring:boot-test"
 include "spock-spring:boot2-test"
