+++ spock-gradle/gradle.gradle
@@ -1,3 +1,12 @@
-dependencies {
-  implementation gradleApi()
+plugins {
+  id 'groovy-gradle-plugin'
+}
+
+gradlePlugin {
+  plugins {
+    spockBasePlugin {
+      id = 'org.spockframework.base'
+      implementationClass = 'org.spockframework.gradle.SpockBasePlugin'
+    }
+  }
 }
