+++ spock-gradle/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'spock-gradle'
+rootProject.buildFileName = 'gradle.gradle'
