--- spock-gradle/src/main/resources/META-INF/gradle-plugins/spock-base.properties
@@ -1,1 +1,0 @@
-implementation-class=org.spockframework.gradle.SpockBasePlugin
